third_party/blink/web_tests/external/wpt/eventsource/eventsource-constructor-empty-url.html

<!doctype html>
<meta charset=utf-8>
<title></title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>

test(function() {
  let es = new EventSource("");
  assert_equals(es.url, document.documentURI,
                "Passing empty string to the EventSource constructor shouldn't throw.");
  es.close();
}, "Passing empty string to the EventSource constructor");

</script>
